refactor: 비속어의 경우 일반 하위 도메인 패키지로 이동
